backend/requirements.txt
aio-pika==9.5.4
aioredis==2.0.1
aiormq==6.8.1
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
async-timeout==5.0.1
asyncpg==0.30.0
awscli==1.37.17
base==0.0.0
billiard==4.2.1
boto3==1.36.16
botocore==1.36.17
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
debugpy==1.8.12
decorator==5.1.1
dnspython==2.7.0
docker==7.1.0
docutils==0.16
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.7
fastapi-cli==0.0.7
flower==2.0.1
fpdf==1.7.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
humanize==4.11.0
idna==3.10
iniconfig==2.0.0
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
kombu==5.4.2
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
packaging==24.2
pamqp==3.3.0
parso==0.8.4
pexpect==4.9.0
pika==1.3.2
pillow==11.1.0
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
psutil==6.1.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pytest==8.3.4
pytest-celery==1.1.3
pytest-docker-tools==3.1.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzbar==0.1.9
qrcode==8.0
rabbitmq==0.2.0
razorpay==1.4.2
redis==5.2.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.7.2
s3transfer==0.11.2
sendgrid==6.11.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
stack-data==0.6.3
starkbank-ecdsa==2.2.0
starlette==0.45.3
svgwrite==1.4.3
tenacity==9.0.0
tornado==6.4.2
traitlets==5.14.3
Tree==0.2.4
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
yarl==1.18.3
